requirements.txt
faker
python-docx
xlsxwriter
fpdf
pandas
matplotlib